API Corrections
Previous pull request had a bug in the creation of event listeners and it was creating duplicate listeners as time went on.  While finding the best way to address this with minimal impact to the code, I trimmed changes to more closely reflect the original code with simple API updates.  Changed mailNotifier.nofication from null to '' in order to use it to track the state of the notification as was originally intended.  Simplified the clear function from previous version.